proc: move from libcyrus_imap to libcyrus_min
will be needed for master to use its functions
