+++ ptclient/ptloader.c
@@ -59,8 +59,8 @@
 #include "auth_pts.h"
 #include "cyrusdb.h"
 #include "imap/global.h"
-#include "imap/proc.h"
 #include "libconfig.h"
+#include "proc.h"
 #include "retry.h"
 #include "xmalloc.h"
 #include "ptloader.h"
